requirements.txt
aiogram==2.23.1
asyncpg==0.27.0
python-dotenv
configparser